Move mock into its own module
